+++ CHANGELOG.rst
@@ -1,7 +1,16 @@
 ********************
-[0.1.5] - 2019-09-25
+[0.2.0] - XXXX-XX-XX
 ********************
 
+**Breaking changes**:
+
+- The ancestors tree sequence now contains the real alleles and not
+  0/1 values as before.
+
+
+********************
+[0.1.5] - 2019-09-25
+********************
 
 **Breaking changes**:
 
@@ -24,7 +33,7 @@
 
 - Change ``tsinfer.UNKNOWN_ALLELE`` to ``tskit.MISSING_DATA`` for marking unknown regions
   of ancestral haplotypes (#188) . This also involves changing the allele storage to a
-  signed int from ``np.uint8`` which matches the tskit v0.2 format for allele storage 
+  signed int from ``np.uint8`` which matches the tskit v0.2 format for allele storage
   (see https://github.com/tskit-dev/tskit/issues/144). Together with the 2 changes above,
   this addition bumped the file format to 3.0.
 
@@ -33,7 +42,7 @@
 - Map non-inference sites onto the tree by using the built-in tskit
   ``map_mutatations`` method. With further work, this should allow triallelic sites
   to be mapped (#185)
-  
+
 - The default tree sequence returned after inference when ``simplify=True`` retains
   unary nodes (i.e. simplify is done with ``keep_unary=True``. This tends to result
   in better compression.